# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple backend project with crud , hashing , authentication and jwt tokens .",
        "version": "1.0.0",
        "title": "CRUD app",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "glorineza2@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:5000",
    "basePath": "/app",
    "tags": [
        {
            "name": "user",
            "description": "Our users",
            "externalDocs": {
                "description": "Find out more",
                "url": "http://swagger.io"
            }
        },
        {
            "name": "auth",
            "description": "Authentication and authorisation apis"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/app": {
            "post": {
                "tags": [
                    "signup"
                ],
                "summary": "Signup to enter in our community",
                "description": "Signup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                    "in": "body",
                    "name": "body",
                    "description": "A new user to be added to the database",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Course"
                    }
                }]
            },
            "put": {
                "tags": [
                    "course"
                ],
                "summary": "Update existing course",
                "description": "Update existing course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Course object that needs to be added to the database",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Course"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "course"
                ],
                "security":[
                    {
                        "Bearer": []
                    }
                ],
                "summary": "Finds all courses",
                "description": "All courses available in the system",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Course"
                            }
                        }
                    },
                    "404": {
                        "description": "No record found"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "course"
                ],
                "summary": "Find course by ID",
                "description": "Returns a single course",
                "operationId": "getCourseById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of course to return",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Course"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Save new user",
                "description": "Save new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "User object that needs to be added to the database",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Finds all users",
                "description": "All Users available in the system",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "404": {
                        "description": "No record found"
                    }
                }
            }
        },
        "/api/users/{email}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Find user by email",
                "description": "Returns a single user",
                "operationId": "getUserByEmail",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "email",
                    "in": "path",
                    "description": "email of course to return",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid email supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/auth/jwt": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login with jwt enabled",
                "description": "Login with jwt enabled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Login object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Login"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/bcrypt": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login with bcrypt enabled",
                "description": "Login with bcrypt enabled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Login object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Login"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "ApiResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "Course": {
            "type": "object",
            "required": [
                "courseName",
                "courseId"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "courseName": {
                    "type": "string",
                    "example": "Javascript"
                },
                "courseId": {
                    "type": "string",
                    "example": "JS001"
                },
                "courseDuration": {
                    "type": "string",
                    "example": "2hr"
                },
                "author": {
                    "type": "string",
                    "example": "James"
                },
                "courseFee": {
                    "type": "string",
                    "example": "100"
                },
                "discount": {
                    "type": "number",
                    "example": "100"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                }
            }
        },
        "Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}